+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ new version (on deck)
 - [ENH] Added `also` method for running functions in chain with no return values.
 - [DOC] Added a ``timeseries`` module section to website docs. Issue #742. @loganthomas
 - [ENH] Added a `pivot_longer` function, a wrapper around ``pd.melt`` and similar to 
-  tidyr's ``pivot_longer`` function. @samukweku
+  tidyr's ``pivot_longer`` function. Also added an example notebook. @samukweku
 - [ENH] Fixed code to returns error if `fill_value` is not a dictionary. @samukweku
 - [INF] Welcome bot (.github/config.yml) for new users added. Issue #739. @samukweku