+++ CHANGELOG.rst
@@ -1,16 +1,21 @@
 new version (on deck)
 =====================
 - [ENH] Added function ``sort_timestamps_monotonically`` to timeseries functions @UGuntupalli
-- [ENH] Added the complete function for converting implicit missing values 
+- [ENH] Added the complete function for converting implicit missing values
   to explicit ones. @samukweku
 - [ENH] Further simplification of expand_grid. @samukweku
 - [BUGFIX] Added copy() method to original dataframe, to avoid mutation. Issue #729. @samukweku
+- [ENH] Added `also` method for running functions in chain with no return values.
+- [DOC] Added a ``timeseries`` module section to website docs. Issue #742. @loganthomas
+- [ENH] Added a `pivot_longer` function, a wrapper around ``pd.melt`` and similar to 
+  tidyr's ``pivot_longer`` function. Also added an example notebook. @samukweku
+- [ENH] Fixed code to returns error if `fill_value` is not a dictionary. @samukweku
+- [INF] Welcome bot (.github/config.yml) for new users added. Issue #739. @samukweku
 
 v0.20.9
 =======
 - [ENH] Updated groupby_agg function to account for null entries in the ``by`` argument. @samukweku
 - [ENH] Added function ``groupby_topk`` to janitor functions @mphirke
-
 
 v0.20.8
 =======