+++ CHANGELOG.rst
@@ -7,10 +7,11 @@ new version (on deck)
 - [BUGFIX] Added copy() method to original dataframe, to avoid mutation. Issue #729. @samukweku
 - [ENH] Added `also` method for running functions in chain with no return values.
 - [DOC] Added a ``timeseries`` module section to website docs. Issue #742. @loganthomas
-- [ENH] Added a `pivot_longer` function, a wrapper around ``pd.melt`` and similar to 
+- [ENH] Added a `pivot_longer` function, a wrapper around ``pd.melt`` and similar to
   tidyr's ``pivot_longer`` function. Also added an example notebook. @samukweku
 - [ENH] Fixed code to returns error if `fill_value` is not a dictionary. @samukweku
 - [INF] Welcome bot (.github/config.yml) for new users added. Issue #739. @samukweku
+- [INF] Deploy ``darglint`` package for docstring linting. Issue #745. @loganthomas
 
 v0.20.9
 =======